requirements.txt
fastapi
uvicorn
ibm_watsonx_ai
python-dotenv